+++ stable_baselines3/common/evaluation.py
@@ -22,11 +22,13 @@ def evaluate_policy(
     Runs policy for ``n_eval_episodes`` episodes and returns average reward.
     This is made to work only with one env.
 
-    If environment has not been wrapped with ``Monitor`` wrapper, reward and
-    episode lengths are counted as it appears with ``env.step`` calls. However
-    if the environment contains wrappers that modify rewards or episode lengths
-    (e.g. reward scaling, early episode reset), these will affect the evaluation
-    results as well.
+    .. note::
+        If environment has not been wrapped with ``Monitor`` wrapper, reward and
+        episode lengths are counted as it appears with ``env.step`` calls. If
+        the environment contains wrappers that modify rewards or episode lengths
+        (e.g. reward scaling, early episode reset), these will affect the evaluation
+        results as well. You can avoid this by wrapping environment with ``Monitor``
+        wrapper before anything else.
 
     :param model: The RL agent you want to evaluate.
     :param env: The gym environment. In the case of a ``VecEnv``